chore: split secret manager file
